fix: renamed organization team model file
