+++ packages/contracts/src/invite.model.ts
@@ -6,7 +6,7 @@ import { IOrganization } from './organization.model';
 import { IOrganizationContact } from './organization-contact.model';
 import { IOrganizationDepartment } from './organization-department.model';
 import { IEmployee } from './employee.model';
-import { IOrganizationTeam } from './organization-team-model';
+import { IOrganizationTeam } from './organization-team.model';
 
 export interface IInvite extends IBasePerTenantAndOrganizationEntityModel {
 	token: string;
@@ -151,8 +151,8 @@ export interface IInviteTeamMemberModel extends IBasePerTenantAndOrganizationEnt
 }
 
 export interface IJoinEmployeeModel extends IBasePerTenantAndOrganizationEntityModel {
-    email: string;
-    employee: IEmployee;
-    organization: IOrganization;
-    languageCode: LanguagesEnum;
+	email: string;
+	employee: IEmployee;
+	organization: IOrganization;
+	languageCode: LanguagesEnum;
 }
