+++ packages/contracts/src/request-approval-team.model.ts
@@ -1,6 +1,6 @@
 import { IBasePerTenantAndOrganizationEntityModel } from './base-entity.model';
 import { IRequestApproval } from './request-approval.model';
-import { IOrganizationTeam } from './organization-team-model';
+import { IOrganizationTeam } from './organization-team.model';
 
 export interface IRequestApprovalTeam
 	extends IBasePerTenantAndOrganizationEntityModel {
