+++ packages/contracts/src/request-approval.model.ts
@@ -1,7 +1,7 @@
 import { IBasePerTenantAndOrganizationEntityModel } from './base-entity.model';
 import { IRequestApprovalEmployee } from './request-approval-employee.model';
 import { IEmployee } from './employee.model';
-import { IOrganizationTeam } from './organization-team-model';
+import { IOrganizationTeam } from './organization-team.model';
 import { IRequestApprovalTeam } from './request-approval-team.model';
 import { IApprovalPolicy } from './approval-policy.model';
 import { ITag } from './tag-entity.model';
